+++ dynamoDBCrud/src/main/java/com/util/dynamodb/CRUDOperations.java
@@ -1,4 +1,4 @@
-package com.amazonaws.samples;
+package com.util.dynamodb;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
